table-construct-backend/requirements.txt
fastapi>=0.68.0,<1.0.0
uvicorn[standard]>=0.15.0,<1.0.0
python-multipart>=0.0.5
chromadb
python-docx
sentence-transformers
numpy
